requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
mediapipe==0.10.14
opencv-python==4.10.0.84
ultralytics>=8.0.0
numpy==1.26.4
yt-dlp>=2024.1.1
httpx>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.0.0
supermemory>=3.27.0
